Problem: cannot get/set device state
slurp by itself is not stateful and is wrapped
for more complex behavior such as slurpif. Since
machinic environments are somewhat isolated by
the host:ip suffix it makes sense to have a simple
stored state for devices that can be called with
a command such as slurpstate or slurps.

Solution: add getting and setting of a device state
that can be accessed and applied before slurp. Also
store and load settings from glworb fields based on
a settings prefix.